SDK-80 add more tests; move test to separate files
