+++ src/validation/api/add-attachment-response.ts
@@ -1,0 +1,9 @@
+import * as t from 'io-ts';
+
+const AddAttachmentResponseIO = t.unknown;
+type AddAttachmentResponse = unknown;
+
+export {
+  AddAttachmentResponseIO,
+  AddAttachmentResponse,
+};
