+++ src/validation/api/upsert-attachment-response.ts
@@ -1,0 +1,9 @@
+import * as t from 'io-ts';
+
+const UpsertAttachmentResponseIO = t.unknown;
+type UpsertAttachmentResponse = unknown;
+
+export {
+  UpsertAttachmentResponseIO,
+  UpsertAttachmentResponse,
+};
